Combined mint to and for
